+++ sdk/storage/azure_storage_blob/src/models/extensions.rs
@@ -13,7 +13,7 @@ use std::collections::HashMap;
 ///
 /// * `self` - The options bag to be modified.
 impl PageBlobClientCreateOptions<'_> {
-    pub fn with_if_not_exists(self) -> Self {
+    pub fn if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -26,7 +26,7 @@ impl PageBlobClientCreateOptions<'_> {
 ///
 /// * `self` - The options bag to be modified.
 impl AppendBlobClientCreateOptions<'_> {
-    pub fn with_if_not_exists(self) -> Self {
+    pub fn if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -39,7 +39,7 @@ impl AppendBlobClientCreateOptions<'_> {
 ///
 /// * `self` - The options bag to be modified.
 impl BlockBlobClientUploadBlobFromUrlOptions<'_> {
-    pub fn with_if_not_exists(self) -> Self {
+    pub fn if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -52,7 +52,7 @@ impl BlockBlobClientUploadBlobFromUrlOptions<'_> {
 ///
 /// * `self` - The options bag to be modified.
 impl BlockBlobClientUploadOptions<'_> {
-    pub fn with_if_not_exists(self) -> Self {
+    pub fn if_not_exists(self) -> Self {
         Self {
             if_none_match: Some("*".into()),
             ..self
@@ -66,7 +66,7 @@ impl BlockBlobClientUploadOptions<'_> {
 /// * `self` - The options bag to be modified.
 /// * `tags` - A HashMap of key-value pairs representing the blob tags.
 impl BlockBlobClientUploadOptions<'_> {
-    pub fn with_tags(self, tags: HashMap<String, String>) -> Self {
+    pub fn tags(self, tags: HashMap<String, String>) -> Self {
         let tags_string = tags
             .iter()
             .map(|(key, value)| format!("{}={}", key, value))