+++ sdk/storage/azure_storage_blob/tests/blob_service_client.rs
@@ -193,7 +193,7 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
         Some(RequestContent::from("hello world".as_bytes().into())),
         Some(
             BlockBlobClientUploadOptions::default()
-                .with_tags(HashMap::from([("foo".to_string(), "bar".to_string())])),
+                .tags(HashMap::from([("foo".to_string(), "bar".to_string())])),
         ),
     )
     .await?;
@@ -203,7 +203,7 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
         Some(RequestContent::from("ferris the crab".as_bytes().into())),
         Some(
             BlockBlobClientUploadOptions::default()
-                .with_tags(HashMap::from([("fizz".to_string(), "buzz".to_string())])),
+                .tags(HashMap::from([("fizz".to_string(), "buzz".to_string())])),
         ),
     )
     .await?;
@@ -212,7 +212,7 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
     create_test_blob(
         &container_client_1.blob_client(&blob3_name.clone()),
         Some(RequestContent::from("six seven".as_bytes().into())),
-        Some(BlockBlobClientUploadOptions::default().with_tags(blob3_tags.clone())),
+        Some(BlockBlobClientUploadOptions::default().tags(blob3_tags.clone())),
     )
     .await?;
 
@@ -224,11 +224,13 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
     }
 
     // Find "hello world" blob by its tag {"foo": "bar"}
-    let response = service_client
-        .find_blobs_by_tags("\"foo\"='bar'", None)
-        .await?;
+    // TODO: Change to find_blobs_by_tags once the emitter support drops.
+    let mut pager = service_client
+        .list_find_blobs_by_tags("\"foo\"='bar'", None)?
+        .into_pages();
+    let response = pager.try_next().await?.unwrap();
     let filter_blob_segment = response.into_model()?;
-    let blobs = filter_blob_segment.blobs.unwrap_or_default();
+    let blobs = filter_blob_segment.blobs;
     assert!(
         blobs
             .iter()
@@ -237,11 +239,13 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
     );
 
     // Find "ferris the crab" blob by its tag {"fizz": "buzz"}
-    let response = service_client
-        .find_blobs_by_tags("\"fizz\"='buzz'", None)
-        .await?;
+    // TODO: Change to find_blobs_by_tags once the emitter support drops.
+    let mut pager = service_client
+        .list_find_blobs_by_tags("\"fizz\"='buzz'", None)?
+        .into_pages();
+    let response = pager.try_next().await?.unwrap();
     let filter_blob_segment = response.into_model()?;
-    let blobs = filter_blob_segment.blobs.unwrap_or_default();
+    let blobs = filter_blob_segment.blobs;
     assert!(
         blobs
             .iter()
@@ -250,11 +254,13 @@ async fn test_find_blobs_by_tags_service(ctx: TestContext) -> Result<(), Box<dyn
     );
 
     // Find "six seven" blob by its tag {"tagged": "true"}
-    let response = service_client
-        .find_blobs_by_tags(&format_filter_expression(&blob3_tags)?, None)
-        .await?;
+    // TODO: Change to find_blobs_by_tags once the emitter support drops.
+    let mut pager = service_client
+        .list_find_blobs_by_tags(&format_filter_expression(&blob3_tags)?, None)?
+        .into_pages();
+    let response = pager.try_next().await?.unwrap();
     let filter_blob_segment = response.into_model()?;
-    let blobs = filter_blob_segment.blobs.unwrap_or_default();
+    let blobs = filter_blob_segment.blobs;
     assert!(
         blobs
             .iter()